requirements.txt
streamlit
numpy
tensorflow
opencv-python-headless
scikit-learn
flask
matplotlib
seaborn
fastapi
pillow
python-multipart
uvicorn
gradio   # for the VTO UI
# transformers